+++ src/main/java/se/iths/clothdatabase/controller/ProductController.java
@@ -4,7 +4,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import se.iths.clothdatabase.entity.ProductEntity;
-import se.iths.clothdatabase.exception.PriceIsMoreThanZeroException;
+import se.iths.clothdatabase.exception.product.PriceIsMoreThanZeroException;
 import se.iths.clothdatabase.service.ProductService;
 
 import java.util.Optional;
