+++ src/main/java/se/iths/clothdatabase/exception/address/IncorrectZIPCodeException.java
@@ -1,4 +1,4 @@
-package se.iths.clothdatabase.exception;
+package se.iths.clothdatabase.exception.address;
 
 public class IncorrectZIPCodeException extends Exception{
     public IncorrectZIPCodeException(String message) {
