+++ src/main/java/se/iths/clothdatabase/exception/userDetails/YoungerThan15Exception.java
@@ -1,4 +1,4 @@
-package se.iths.clothdatabase.exception;
+package se.iths.clothdatabase.exception.userDetails;
 
 public class YoungerThan15Exception extends Exception{
     public YoungerThan15Exception(String message) {
