+++ src/main/java/se/iths/clothdatabase/service/UserService.java
@@ -6,9 +6,9 @@ import se.iths.clothdatabase.entity.PaymentEntity;
 import se.iths.clothdatabase.entity.ProductEntity;
 import se.iths.clothdatabase.entity.RoleEntity;
 import se.iths.clothdatabase.entity.UserEntity;
-import se.iths.clothdatabase.exception.LessThanThreeCharacterException;
-import se.iths.clothdatabase.exception.NotEnoughMoneyException;
-import se.iths.clothdatabase.exception.ProductIsNotInStockException;
+import se.iths.clothdatabase.exception.user.LessThanThreeCharacterException;
+import se.iths.clothdatabase.exception.user.NotEnoughMoneyException;
+import se.iths.clothdatabase.exception.product.ProductIsNotInStockException;
 import se.iths.clothdatabase.repository.PaymentRepository;
 import se.iths.clothdatabase.repository.ProductRepository;
 import se.iths.clothdatabase.repository.RoleRepository;
