+++ src/test/java/se/iths/clothdatabase/exception/ProductExceptionsTest.java
@@ -5,6 +5,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.TestPropertySource;
 import se.iths.clothdatabase.entity.ProductEntity;
+import se.iths.clothdatabase.exception.product.PriceIsMoreThanZeroException;
 import se.iths.clothdatabase.service.ProductService;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
