+++ src/test/java/se/iths/clothdatabase/exception/UserExceptionsTest.java
@@ -5,6 +5,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.TestPropertySource;
 import se.iths.clothdatabase.entity.UserEntity;
+import se.iths.clothdatabase.exception.user.LessThanThreeCharacterException;
 import se.iths.clothdatabase.service.UserService;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
